requirements.txt
jupyter
tensorflow-gpu
matplotlib
